Big refactoring for better project structure
