+++ src/main/java/home/andreybelonog/model/repository/BookRepository.java
@@ -1,6 +1,6 @@
-package home.andreybelonog.repository;
+package home.andreybelonog.model.repository;
 
-import home.andreybelonog.domain.Book;
+import home.andreybelonog.model.entity.Book;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BookRepository extends CrudRepository<Book, Long> {
