+++ src/main/java/home/andreybelonog/model/service/BookService.java
@@ -1,0 +1,4 @@
+package home.andreybelonog.model.service;
+
+public interface BookService {
+}
