+++ src/main/java/home/andreybelonog/model/service/impl/BookServiceImpl.java
@@ -1,0 +1,4 @@
+package home.andreybelonog.model.service.impl;
+
+public class BookServiceImpl {
+}
